orchestrator/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Docker
docker==7.1.0
kubernetes==31.0.0

# Utils
httpx==0.27.2
python-dotenv==1.0.1
